requirements.txt
python-dotenv
python-Levenshtein
python-telegram-bot
pytz
PyYAML
requests
thefuzz[speedup]
